LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/srcdoc_process_attributes-expected.txt



PASS Adding `srcdoc` attribute triggers attributes processing 
PASS Setting `srcdoc` (via property) triggers attributes processing 
FAIL Removing `srcdoc` attribute triggers attributes processing assert_equals: expected "blob:" but got "about:"
